refactor adminLogin sc-5676
