+++ step_definitions/adminLogin-steps.js
@@ -6,6 +6,7 @@ const loginData = require('../shared-objects/loginData');
 const shared = { loginData };
 const firstLogin = require('../shared_steps/firstLogin.js');
 const elementHelpers = require('../runtime/helpers/elementHelpers.js');
+const DashboardPage = require('../page-objects/pages/DashboardPage');
 
 Given(/^an admin arrives on the Schul-Cloud login homepage$/, function() {
 	return elementHelpers.loadPage(shared.loginData.url, 10);
@@ -33,18 +34,18 @@ Then(
 Then(
 	/^the admin-dashboard should have the correct school$/,
 	function() {
-		return adminLogin.loginSchool();
+		return DashboardPage.loginSchool();
 	}
 );
 
 Then(
 	/^the admin-dashboard should have the admin name and profession$/,
 	function() {
-		return adminLogin.loginFullUserInfo();
+		return DashboardPage.loginFullUserInfo();
 	}
 );
 
 Then(/^the admin-dashboard should have the following tabs$/, function (LoginTextLabels) {
-    return adminLogin.checkIfElementIsVisisble(LoginTextLabels, shared.loginData.elem.loginTabs);
+    return DashboardPage.checkIfTabsAreVisible(LoginTextLabels, shared.loginData.elem.loginTabs);
 	}
 );
