+++ step_definitions/visibilityOfDialoguesTabsForTheTeacher-steps.js
@@ -7,6 +7,7 @@ const shared = { loginData };
 const administration = require('../page-objects/administration');
 const firstLogin = require('../shared_steps/firstLogin.js');
 const elementHelpers = require('../runtime/helpers/elementHelpers.js');
+const DashboardPage = require('../page-objects/pages/DashboardPage');
 
 
   Given(/^The Teacher arrives on the Schul-Cloud page$/, function () {
@@ -28,5 +29,5 @@ const elementHelpers = require('../runtime/helpers/elementHelpers.js');
   });
 
   Then(/^Verify if all required tabs are visible in Verwaltung area$/, function (administrationTextLabels) {
-    return administration.checkIfElementIsVisisble(administrationTextLabels, Admin.administrationsTabs);
+    return DashboardPage.checkIfTabsAreVisible(administrationTextLabels, Admin.administrationsTabs);
     });
